requirement.txt
asgiref==3.2.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
